backend/pyproject.toml
[project]
name = "receptionist-ai-backend"
version = "0.1.0"
description = "AI Receptionist Backend - Keystone"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.5.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pydantic>=2.0.0",
    "openai>=1.50.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
